backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
scipy
pillow
python-multipart

